Improve benchmark differential output

+++ src/bench/benchmark.scala
@@ -58,6 +58,6 @@ extension [TestType](test: Test[TestType])
     val mean: Double = total.toDouble/count
     val max: Long = times.max
     val variance: Double = (times.map { t => (mean - t)*(mean - t) }.sum)/count
-    val stdDev: Double = math.sqrt(variance.toDouble)
+    val stdDev: Double = math.sqrt(variance)
     val benchmark = Benchmark(total, times.size, min, mean, max, stdDev, confidence.or(95), baseline)
     inc.include(runner.report, test.id, benchmark)
